Fix 1.0

+++ src/main/java/com/github/eternajunior/model/Account.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Account {
-    private int id;
+    private Long id;
     private AccountStatus accountStatus;
     private String name;
-    private int balance;
+    private Integer balance;
 }
